@@ -1,6 +1,6 @@
 {
   "name": "hoist-non-react-statics",
-  "version": "1.0.3",
+  "version": "1.0.5",
   "description": "Copies non-react specific statics from a child component to a parent component",
   "main": "index.js",
   "repository": {
@@ -25,27 +25,27 @@
   "devDependencies": {
     "babel": "^5.0.7",
     "babel-eslint": "^3.0.1",
-    "chai": "^2.0.0",
+    "chai": "^3.0.0",
     "coveralls": "^2.11.1",
     "eslint": "^0.21.0",
     "istanbul": "^0.3.2",
     "mocha": "^2.0.1",
     "pre-commit": "^1.0.7",
-    "react": "^0.13.3"
+    "react": "<0.15.0"
   },
   "keywords": [
     "react"
   ],
-  "gitHead": "91e65b1c10e3a37d5f94ef1424b3233e442d69e9",
+  "gitHead": "24395322f31203484f3cfb6958c2b4313006483f",
   "bugs": {
     "url": "https://github.com/mridgway/hoist-non-react-statics/issues"
   },
   "homepage": "https://github.com/mridgway/hoist-non-react-statics#readme",
-  "_id": "hoist-non-react-statics@1.0.3",
-  "_shasum": "22cfc787e53cc7cf4832ee35214fa3c55ef2d8aa",
+  "_id": "hoist-non-react-statics@1.0.5",
+  "_shasum": "0e36d2c130c8511f267a0d4ceb45ec7d7e4f0c70",
   "_from": "hoist-non-react-statics@>=1.0.3 <2.0.0",
-  "_npmVersion": "2.11.3",
-  "_nodeVersion": "0.10.33",
+  "_npmVersion": "2.14.15",
+  "_nodeVersion": "0.10.40",
   "_npmUser": {
     "name": "mridgway",
     "email": "mcridgway@gmail.com"
@@ -57,9 +57,13 @@
     }
   ],
   "dist": {
-    "shasum": "22cfc787e53cc7cf4832ee35214fa3c55ef2d8aa",
-    "tarball": "http://registry.npmjs.org/hoist-non-react-statics/-/hoist-non-react-statics-1.0.3.tgz"
+    "shasum": "0e36d2c130c8511f267a0d4ceb45ec7d7e4f0c70",
+    "tarball": "http://registry.npmjs.org/hoist-non-react-statics/-/hoist-non-react-statics-1.0.5.tgz"
+  },
+  "_npmOperationalInternal": {
+    "host": "packages-5-east.internal.npmjs.com",
+    "tmp": "tmp/hoist-non-react-statics-1.0.5.tgz_1454447620415_0.7366878658067435"
   },
   "directories": {},
-  "_resolved": "https://registry.npmjs.org/hoist-non-react-statics/-/hoist-non-react-statics-1.0.3.tgz"
+  "_resolved": "https://registry.npmjs.org/hoist-non-react-statics/-/hoist-non-react-statics-1.0.5.tgz"
 }
