@@ -1,31 +1,27 @@
-const { Component, PropTypes, Children } = require('react')
-const storeShape = require('../utils/storeShape')
+import { Component, PropTypes, Children } from 'react'
+import storeShape from '../utils/storeShape'
 
-if (process.env.NODE_ENV !== 'production') {
-  let didWarnAboutReceivingStore = false
-  /* eslint-disable no-var */
-  var warnAboutReceivingStore = function () {
-    /* eslint-enable no-var */
-    if (didWarnAboutReceivingStore) {
-      return
-    }
-    didWarnAboutReceivingStore = true
-
-    /* eslint-disable no-console */
-    if (typeof console !== 'undefined' && typeof console.error === 'function') {
-      console.error(
-        '<Provider> does not support changing `store` on the fly. ' +
-        'It is most likely that you see this error because you updated to ' +
-        'Redux 2.x and React Redux 2.x which no longer hot reload reducers ' +
-        'automatically. See https://github.com/rackt/react-redux/releases/' +
-        'tag/v2.0.0 for the migration instructions.'
-      )
-    }
-    /* eslint-disable no-console */
+let didWarnAboutReceivingStore = false
+function warnAboutReceivingStore() {
+  if (didWarnAboutReceivingStore) {
+    return
   }
+  didWarnAboutReceivingStore = true
+
+  /* eslint-disable no-console */
+  if (typeof console !== 'undefined' && typeof console.error === 'function') {
+    console.error(
+      '<Provider> does not support changing `store` on the fly. ' +
+      'It is most likely that you see this error because you updated to ' +
+      'Redux 2.x and React Redux 2.x which no longer hot reload reducers ' +
+      'automatically. See https://github.com/rackt/react-redux/releases/' +
+      'tag/v2.0.0 for the migration instructions.'
+    )
+  }
+  /* eslint-disable no-console */
 }
 
-class Provider extends Component {
+export default class Provider extends Component {
   getChildContext() {
     return { store: this.store }
   }
@@ -59,5 +55,3 @@ Provider.propTypes = {
 Provider.childContextTypes = {
   store: storeShape.isRequired
 }
-
-module.exports = Provider
