@@ -1,9 +1,7 @@
-const { PropTypes } = require('react')
+import { PropTypes } from 'react'
 
-const storeShape = PropTypes.shape({
+export default PropTypes.shape({
   subscribe: PropTypes.func.isRequired,
   dispatch: PropTypes.func.isRequired,
   getState: PropTypes.func.isRequired
 })
-
-module.exports = storeShape
