@@ -1,7 +1,5 @@
 import { bindActionCreators } from 'redux'
 
-function wrapActionCreators(actionCreators) {
+export default function wrapActionCreators(actionCreators) {
   return dispatch => bindActionCreators(actionCreators, dispatch)
 }
-
-module.exports = wrapActionCreators
