@@ -1,14 +1,13 @@
+'use strict';
+
+exports.__esModule = true;
+exports["default"] = warning;
 /**
  * Prints a warning in the console if it exists.
  *
  * @param {String} message The warning message.
  * @returns {void}
  */
-'use strict';
-
-exports.__esModule = true;
-exports['default'] = warning;
-
 function warning(message) {
   /* eslint-disable no-console */
   if (typeof console !== 'undefined' && typeof console.error === 'function') {
@@ -23,5 +22,3 @@ function warning(message) {
   } catch (e) {}
   /* eslint-enable no-empty */
 }
-
-module.exports = exports['default'];
